+++ WorkoutMirrorLive/FeaturesNew/ActivitiesTab/Feature/ActivitiesFeature+Destination.swift
@@ -1,0 +1,18 @@
+
+//
+//  ActivitiesFeature+Destination.swift
+//  MyFitnessJournal
+//
+//  Created by Sebastian Sciuba on 15/12/2025.
+//
+
+import ComposableArchitecture
+
+/// Implementation of `ActivitiesFeature` destination
+extension ActivitiesFeature {
+    
+    @Reducer
+    enum Destination {
+        
+    }
+}
